feat(db) groundwork for new migrations framework
* entirely built into new DAO (`kong.db`)
* only Cassandra support for now
* better design for schema_meta table (storing past executed migrations)
* concept of "subsystem": core, plugins, and "EE" are all possible
  subsystem, each with their own set of migrations
* migrations are in separate files, indexed in `kong.db.migrations` for
  the core ones
* migrations are usually split in two steps: non-destructive, and
  destructive steps
* new CLI commands: check, bootstrap, finish (+ up and reset as
  previously)
* migrations run within cluster_mutex() (i.e. leader-election), which
  means they can be launched simultaneously from any number of nodes, and
  not a single one anymore
* migrations can be "executed" or "pending", when their second
  (destructive) step is still pending
* hypothetical support for "translation functions", for breaking changes
  on-the-fly without downtime or incompatibility time windows
* squashed core Cassandra migrations into one

+++ kong/db/strategies/cassandra/connector.lua
@@ -145,12 +145,56 @@ function CassandraConnector:connect()
 end
 
 
+-- open a connection from the first available contact point,
+-- without a keyspace
+function CassandraConnector:connect_migrations(opts)
+  if self.connection then
+    return
+  end
+
+  opts = opts or {}
+
+  local peer, err = self.cluster:first_coordinator()
+  if not peer then
+    return nil, "failed to acquire contact point: " .. err
+  end
+
+  if opts.use_keyspace then
+    local ok, err = peer:change_keyspace(self.keyspace)
+    if not ok then
+      return nil, err
+    end
+  end
+
+  self.connection = peer
+
+  return true
+end
+
+
 function CassandraConnector:setkeepalive()
   if not self.connection then
     return
   end
 
   local ok, err = self.connection:setkeepalive()
+
+  self.connection = nil
+
+  if not ok then
+    return nil, err
+  end
+
+  return true
+end
+
+
+function CassandraConnector:close()
+  if not self.connection then
+    return
+  end
+
+  local ok, err = self.connection:close()
 
   self.connection = nil
 
@@ -204,7 +248,35 @@ function CassandraConnector:query(query, args, opts, operation)
 end
 
 
+local function select_keyspaces(self)
+  if not self.connection then
+    error("no connection")
+  end
+
+  if not self.major_version then
+    return nil, "missing self.major_version"
+  end
+
+  local cql
+
+  if self.major_version == 3 then
+    cql = [[SELECT * FROM system_schema.keyspaces
+              WHERE keyspace_name = ?]]
+
+  else
+    cql = [[SELECT * FROM system.schema_keyspaces
+              WHERE keyspace_name = ?]]
+  end
+
+  return self.connection:execute(cql, { self.keyspace })
+end
+
+
 local function select_tables(self)
+  if not self.connection then
+    error("no connection")
+  end
+
   if not self.major_version then
     return nil, "missing self.major_version"
   end
@@ -219,7 +291,7 @@ local function select_tables(self)
             WHERE keyspace_name = ?]]
   end
 
-  return self:query(cql, { self.keyspace })
+  return self.connection:execute(cql, { self.keyspace })
 end
 
 
@@ -280,7 +352,9 @@ function CassandraConnector:truncate()
                        and rows[i].table_name
                        or rows[i].columnfamily_name
 
-    if table_name ~= "schema_migrations" then
+    if table_name ~= "schema_migrations" and
+       table_name ~= "schema_meta" and
+       table_name ~= "locks" then
       local cql = string.format("TRUNCATE TABLE %s.%s",
                                 self.keyspace, table_name)
 
@@ -389,6 +463,279 @@ function CassandraConnector:remove_lock(key, owner)
   end
 
   return true
+end
+
+
+do
+  -- migrations
+  local pl_stringx = require "pl.stringx"
+
+
+  local SCHEMA_META_KEY = "schema_meta"
+
+
+  function CassandraConnector:schema_migrations()
+    if not self.connection then
+      error("no connection")
+    end
+
+    do
+      -- has keyspace table?
+
+      local rows, err = select_keyspaces(self)
+      if not rows then
+        return nil, err
+      end
+
+      local has_keyspace
+
+      for _, row in ipairs(rows) do
+        if row.keyspace_name == self.keyspace then
+          has_keyspace = true
+          break
+        end
+      end
+
+      if not has_keyspace then
+        -- no keyspace needs bootstrap
+        return nil
+      end
+    end
+
+    do
+      -- has schema_meta table?
+
+      local rows, err = select_tables(self)
+      if not rows then
+        return nil, err
+      end
+
+      local has_schema_meta_table
+
+      for _, row in ipairs(rows) do
+        -- Cassandra 3: table_name
+        -- Cassadra 2: columnfamily_name
+        if row.table_name == "schema_meta"
+          or row.columnfamily_name == "schema_meta" then
+          has_schema_meta_table = true
+          break
+        end
+      end
+
+      if not has_schema_meta_table then
+        -- keyspace, but no schema: needs bootstrap
+        return nil
+      end
+    end
+
+    local ok, err = self.connection:change_keyspace(self.keyspace)
+    if not ok then
+      return nil, err
+    end
+
+    do
+      -- has migrations?
+
+      local rows, err = self.connection:execute([[
+        SELECT * FROM schema_meta WHERE key = ?
+      ]], {
+        SCHEMA_META_KEY,
+      })
+      if not rows then
+        return nil, err
+      end
+
+      -- no migrations: is bootstrapped but not migrated
+      -- migrations: has some migrations
+      return rows
+    end
+  end
+
+
+  function CassandraConnector:schema_bootstrap(kong_config)
+    -- compute keyspace creation CQL
+
+    local cql_replication
+    local cass_repl_strategy = kong_config.cassandra_repl_strategy
+
+    if cass_repl_strategy == "SimpleStrategy" then
+      cql_replication = string.format([[
+        {'class': 'SimpleStrategy', 'replication_factor': %d}
+      ]], kong_config.cassandra_repl_factor)
+
+    elseif cass_repl_strategy == "NetworkTopologyStrategy" then
+      local dcs = {}
+
+      for _, dc_conf in ipairs(kong_config.cassandra_data_centers) do
+        local dc_name, dc_repl = string.match(dc_conf, "([^:]+):(%d+)")
+        if dc_name and dc_repl then
+          table.insert(dcs, string.format("'%s': %s", dc_name, dc_repl))
+        end
+      end
+
+      if #dcs < 1 then
+        return nil, "invalid cassandra_data_centers configuration"
+      end
+
+      cql_replication = string.format([[
+        {'class': 'NetworkTopologyStrategy', %s}
+      ]], table.concat(dcs, ", "))
+
+    else
+      error("unknown replication_strategy: " .. tostring(cass_repl_strategy))
+    end
+
+    -- get a contact point connection (no keyspace set)
+
+    if not self.connection then
+      error("no connection")
+    end
+
+    -- create keyspace if not exists
+
+    local res, err = self.connection:execute(string.format([[
+      CREATE KEYSPACE IF NOT EXISTS %q
+      WITH REPLICATION = %s
+    ]], kong_config.cassandra_keyspace, cql_replication))
+    if not res then
+      return nil, err
+    end
+
+    local ok, err = self.connection:change_keyspace(kong_config.cassandra_keyspace)
+    if not ok then
+      return nil, err
+    end
+
+    -- create schema meta table if not exists
+
+    local res, err = self.connection:execute([[
+      CREATE TABLE IF NOT EXISTS schema_meta (
+        key             text,
+        subsystem       text,
+        last_executed   text,
+        executed        set<text>,
+        pending         set<text>,
+
+        PRIMARY KEY (key, subsystem)
+      )
+    ]])
+    if not res then
+      return nil, err
+    end
+
+    ok, err = self.cluster:wait_schema_consensus(self.connection)
+    if not ok then
+      return nil, err
+    end
+
+    return true
+  end
+
+
+  function CassandraConnector:schema_reset()
+    if not self.connection then
+      error("no connection")
+    end
+
+    local ok, err = self.connection:execute(string.format([[
+      DROP KEYSPACE IF EXISTS %q
+    ]], self.keyspace))
+    if not ok then
+      return nil, err
+    end
+
+    ok, err = self.cluster:wait_schema_consensus(self.connection)
+    if not ok then
+      return nil, err
+    end
+
+    return true
+  end
+
+
+  function CassandraConnector:run_up_migration(up_cql)
+    if type(up_cql) ~= "string" then
+      error("up_cql must be a string", 2)
+    end
+
+    if not self.connection then
+      error("no connection")
+    end
+
+    local t_cql = pl_stringx.split(up_cql, ";")
+
+    for i = 1, #t_cql do
+      local cql = pl_stringx.strip(t_cql[i])
+      if cql ~= "" then
+        local res, err = self.connection:execute(cql)
+        if not res then
+          return nil, err
+        end
+      end
+    end
+
+    return true
+  end
+
+
+  function CassandraConnector:record_migration(subsystem, name, state)
+    if type(subsystem) ~= "string" then
+      error("subsystem must be a string", 2)
+    end
+
+    if type(name) ~= "string" then
+      error("name must be a string", 2)
+    end
+
+    if not self.connection then
+      error("no connection")
+    end
+
+    local cql
+    local args
+
+    if state == "executed" then
+      cql = [[UPDATE schema_meta
+              SET last_executed = ?, executed = executed + ?
+              WHERE key = ? AND subsystem = ?]]
+      args = {
+        name,
+        cassandra.set({ name }),
+      }
+
+    elseif state == "pending" then
+      cql = [[UPDATE schema_meta
+              SET pending = pending + ?
+              WHERE key = ? AND subsystem = ?]]
+      args = {
+        cassandra.set({ name }),
+      }
+
+    elseif state == "teardown" then
+      cql = [[UPDATE schema_meta
+              SET pending = pending - ?, executed = executed + ?,
+                  last_executed = ?
+              WHERE key = ? AND subsystem = ?]]
+      args = {
+        cassandra.set({ name }),
+        cassandra.set({ name }),
+        name,
+      }
+
+    else
+      error("unknown 'state' argument: " .. tostring(state))
+    end
+
+    table.insert(args, SCHEMA_META_KEY)
+    table.insert(args, subsystem)
+
+    local res, err = self.connection:execute(cql, args)
+    if not res then
+      return nil, err
+    end
+
+    return true
+  end
 end
